+++ libs/adls/src/thds/adls/__init__.py
@@ -1,8 +1,8 @@
 from thds.core import meta
 
+from .cached_up_down import download_to_cache, upload_through_cache  # noqa: F401
 from .fqn import *  # noqa: F401,F403
 from .impl import *  # noqa: F401,F403
-from .resource import AdlsHashedResource, get_read_only  # noqa: F401
 from .ro_cache import Cache, global_cache  # noqa: F401
 
 __version__ = meta.get_version(__name__)