+++ libs/adls/src/thds/adls/_upload.py
@@ -1,4 +1,7 @@
-import shutil
+"""Just utilities for deciding whether or not to upload.
+
+Not an officially-published API of the thds.adls library.
+"""
 import typing as ty
 from pathlib import Path
 
@@ -6,13 +9,9 @@ import azure.core.exceptions
 from azure.storage.blob import ContentSettings
 from azure.storage.filedatalake import DataLakeFileClient, FileProperties
 
-from thds.core import log, scope
+from thds.core import log
 
-from .fqn import AdlsFqn
-from .global_client import get_global_client
-from .link import link, set_read_only
 from .md5 import AnyStrSrc, try_md5
-from .ro_cache import Cache
 
 _SKIP_ALREADY_UPLOADED_CHECK_IF_MORE_THAN_BYTES = 2 * 2**20  # 2 MB is about right
 
@@ -32,41 +31,61 @@ def _get_checksum_content_settings(data: AnyStrSrc) -> ty.Optional[ContentSettin
     return None
 
 
+def _too_small_to_skip_upload(data: AnyStrSrc, min_size_for_remote_check: int) -> bool:
+    def _len() -> int:
+        if isinstance(data, Path) and data.exists():
+            return data.stat().st_size
+        try:
+            return len(data)  # type: ignore
+        except TypeError as te:
+            logger.debug(f"failed to get length? {repr(te)} for {data}")
+            return min_size_for_remote_check + 1
+
+    return _len() < min_size_for_remote_check
+
+
+class UploadDecision(ty.NamedTuple):
+    upload_required: bool
+    content_settings: ty.Optional[ContentSettings]
+
+
 def _co_content_settings_for_upload_unless_file_present_with_matching_checksum(
-    data: AnyStrSrc,
-) -> ty.Generator[bool, FileProperties, ty.Optional[ContentSettings]]:
+    data: AnyStrSrc, min_size_for_remote_check: int
+) -> ty.Generator[bool, FileProperties, UploadDecision]:
     local_content_settings = _get_checksum_content_settings(data)
     if not local_content_settings:
-        return ContentSettings()  # need to upload but we have no md5 info
-    try:
-        if len(data) < _SKIP_ALREADY_UPLOADED_CHECK_IF_MORE_THAN_BYTES:  # type: ignore
-            logger.debug("Too small to bother with an early call - let's just upload...")
-            return local_content_settings
-    except TypeError as te:
-        logger.debug(f"failed to get length? {repr(te)} for {data}")
+        return UploadDecision(True, None)
+    if _too_small_to_skip_upload(data, min_size_for_remote_check):
+        logger.debug("Too small to bother with an early call - let's just upload...")
+        return UploadDecision(True, local_content_settings)
     props = yield True
     if not props:
         logger.debug("No remote file properties could be fetched so an upload is required")
-        return local_content_settings
+        return UploadDecision(True, local_content_settings)
     if props.content_settings.content_md5 == local_content_settings.content_md5:
         logger.info(f"Remote file {props.name} already exists and has matching checksum")
-        return None
+        return UploadDecision(False, local_content_settings)
     logger.debug("Remote file exists but MD5 does not match - upload required.")
-    return local_content_settings
+    return UploadDecision(True, local_content_settings)
 
 
 doc = """
-Returns ContentSettings if the upload is necessary; otherwise
-returns None which means you can forgo the upload since it's
-already previously occurred.
+Returns False for upload_required if the file is large and the remote
+exists and has a known, matching checksum.
+
+Returns ContentSettings if an MD5 checksum can be calculated.
 """
 
 
-async def async_content_settings_if_upload_required(
-    fc: DataLakeFileClient, data: AnyStrSrc
-) -> ty.Optional[ContentSettings]:
+async def async_upload_decision_and_settings(
+    fc: DataLakeFileClient,
+    data: AnyStrSrc,
+    min_size_for_remote_check: int = _SKIP_ALREADY_UPLOADED_CHECK_IF_MORE_THAN_BYTES,
+) -> UploadDecision:
     try:
-        co = _co_content_settings_for_upload_unless_file_present_with_matching_checksum(data)
+        co = _co_content_settings_for_upload_unless_file_present_with_matching_checksum(
+            data, min_size_for_remote_check
+        )
         while True:
             co.send(None)
             try:
@@ -77,11 +96,15 @@ async def async_content_settings_if_upload_required(
         return stop.value
 
 
-def content_settings_if_upload_required(
-    fc: DataLakeFileClient, data: AnyStrSrc
-) -> ty.Optional[ContentSettings]:
+def upload_decision_and_settings(
+    fc: DataLakeFileClient,
+    data: AnyStrSrc,
+    min_size_for_remote_check: int = _SKIP_ALREADY_UPLOADED_CHECK_IF_MORE_THAN_BYTES,
+) -> UploadDecision:
     try:
-        co = _co_content_settings_for_upload_unless_file_present_with_matching_checksum(data)
+        co = _co_content_settings_for_upload_unless_file_present_with_matching_checksum(
+            data, min_size_for_remote_check
+        )
         while True:
             co.send(None)
             try:
@@ -92,50 +115,5 @@ def content_settings_if_upload_required(
         return stop.value
 
 
-async_content_settings_if_upload_required.__doc__ = doc
-content_settings_if_upload_required.__doc__ = doc
-
-
-UploadSrc = ty.Union[Path, bytes, ty.IO[ty.AnyStr]]
-
-
-def _write_through_local_cache(local_cache_path: Path, data: UploadSrc) -> ty.Optional[Path]:
-    if isinstance(data, Path) and data.exists():
-        if not link(data, local_cache_path):
-            shutil.copyfile(data, local_cache_path)
-        set_read_only(local_cache_path)
-        return local_cache_path
-    if hasattr(data, "read") and hasattr(data, "seek"):
-        with local_cache_path.open("wb") as f:
-            f.write(data.read())  # type: ignore
-            data.seek(0)  # type: ignore
-        set_read_only(local_cache_path)
-        return local_cache_path
-    if isinstance(data, bytes):
-        with local_cache_path.open("wb") as f:
-            f.write(data)
-        set_read_only(local_cache_path)
-        return local_cache_path
-    return None
-
-
-@scope.bound
-def upload(fqn: AdlsFqn, data: UploadSrc, write_through_cache: ty.Optional[Cache] = None) -> None:
-    """Uploads only if the remote does not exist or does not match
-    md5.
-
-    Always embeds md5 in upload.
-
-    Can write through a local cache if provided.
-    """
-    if write_through_cache:
-        data = _write_through_local_cache(write_through_cache.path(fqn), data) or data
-        # we can now upload from the local cache path, which has the
-        # advantage (over some kinds of readable byte iterables) of
-        # guaranteeing that we can extract an md5.
-
-    fs_client = get_global_client(fqn.sa, fqn.container).get_file_client(fqn.path)
-    if content_settings := content_settings_if_upload_required(fs_client, data):
-        if isinstance(data, Path):
-            data = scope.enter(open(data, "rb"))
-        fs_client.upload_data(data, overwrite=True, content_settings=content_settings)
+async_upload_decision_and_settings.__doc__ = doc
+upload_decision_and_settings.__doc__ = doc