+++ libs/adls/src/thds/adls/link.py
@@ -1,10 +1,10 @@
 """Best-effort to link a destination to a source depending on file system support."""
-import contextlib
 import os
 import platform
 import stat
 import subprocess
 import typing as ty
+from pathlib import Path
 
 from thds.core import log
 from thds.core import types as ct
@@ -16,7 +16,11 @@ logger = log.getLogger(__name__)
 LinkType = ty.Literal["same", "ref", "hard", "soft", ""]
 
 
-def link(src: ct.StrOrPath, dest: ct.StrOrPath) -> LinkType:
+def link(
+    src: ct.StrOrPath,
+    dest: ct.StrOrPath,
+    *attempt_types: LinkType,
+) -> LinkType:
     """Attempt reflink, then hardlink, then softlink.
 
     First removes any existing file at dest.
@@ -25,27 +29,39 @@ def link(src: ct.StrOrPath, dest: ct.StrOrPath) -> LinkType:
 
     Return empty string if no link could be created.
     """
-    if str(src) == str(dest):
+    if not attempt_types:
+        attempt_types = ("ref", "hard", "soft")
+    src = Path(src).resolve()
+    if src == Path(dest).resolve():
         return "same"
-
-    with contextlib.suppress(FileNotFoundError):
+    assert os.path.exists(src), f"Source {src} does not exist"
+    try:
         os.remove(dest)  # link will fail if the path already exists
-    if _IS_MAC:
+        logger.warning(f'Removed existing file at "{dest}"')
+    except FileNotFoundError:
+        pass
+    if _IS_MAC and "ref" in attempt_types:
         try:
             subprocess.check_output(["cp", "-c", str(src), str(dest)])
+            logger.info(f"Created a copy-on-write reflink from {src} to {dest}")
             return "ref"
         except subprocess.CalledProcessError:
             pass
-    try:
-        os.link(src, dest)
-        return "hard"
-    except OSError as oserr:
-        logger.warning(f"Unable to hard-link {src} to {dest}" f" ({oserr}); falling back to soft link.")
-    try:
-        os.symlink(src, dest)
-        return "soft"
-    except OSError as oserr:
-        logger.warning(f"Unable to soft-link {src} to {dest}" f" ({oserr})")
+    if "hard" in attempt_types:
+        try:
+            os.link(src, dest)
+            logger.info(f"Created a hardlink from {src} to {dest}")
+            return "hard"
+        except OSError as oserr:
+            logger.warning(f"Unable to hard-link {src} to {dest} ({oserr})")
+    if "soft" in attempt_types:
+        try:
+            os.symlink(src, dest)
+            assert os.path.exists(dest), dest
+            logger.info(f"Created a softlink from {src} to {dest}")
+            return "soft"
+        except OSError as oserr:
+            logger.warning(f"Unable to soft-link {src} to {dest}" f" ({oserr})")
 
     return ""